refactor: Post 및 User 인터페이스 위치 변경
