+++ src/entities/post/model/type.ts
@@ -1,0 +1,11 @@
+interface Post {
+  id: number
+  userId: number
+  title: string
+  content: string
+  createdAt: string
+  updatedAt: string
+  body: string
+}
+
+export default Post
