--- src/features/posts/model/type.ts
@@ -1,11 +1,0 @@
-interface Post {
-  id: number
-  userId: number
-  title: string
-  content: string
-  createdAt: string
-  updatedAt: string
-  body: string
-}
-
-export default Post
